Finish Rock Paper Scissors game
This commit adds and changes multiple things.
Changes return of playRound function to array, because
we need an information who won a round. The array
has a string which contains this information at 0 and
string which might be shown to user at 1. Game function got
created. It allows user to play 5 rounds against computer.
It takes in loop playRound function as well as
user prompt and computer generated output. It shows
result of each round. It is capable of tracking user and
computer score and displays a message at the end, who won
the overall 5 round "Rock, Paper, Scissors" game.